Adds the windowinfo to the design system and the ui themes
