+++ app/build.gradle.kts
@@ -57,6 +57,7 @@ dependencies {
     implementation(project(":data"))
     implementation(project(":domain"))
     implementation(project(":presentation"))
+    implementation(project(":designsystem"))
 
     implementation(libs.androidx.core.splashscreen)
     implementation(libs.koin.core)
