+++ designsystem/src/main/java/me/androidbox/designsystem/ui/theme/Color.kt
@@ -2,26 +2,27 @@ package me.androidbox.designsystem.ui.theme
 
 import androidx.compose.ui.graphics.Color
 
-val BusbyGreen = Color(0xFF00F15E)
+val BusbyGreen = Color(0xFFC4C4C4)
 val BusbyGreen30 = Color(0x4D00F15E)
 val BusbyGreen10 = Color(0x1A00F15E)
 val BusbyGreen5 = Color(0x0D00F15E)
-val BusbyBlack = Color(0xFF080707)
-val BusbyGray = Color(0xFF87938C)
+val BusbyBlack = Color(0xFF333333)
+val BusbyGray = Color(0xFFC4C4C4)
 val BusbyGray40 = Color(0x6687938C)
 val BusbyDarkGray = Color(0xFF232624)
 val BusbyWhite = Color(0xFFFAFAFA)
 val BusbyDarkRed = Color(0xFFBB3D3D)
 val BusbyDarkRed5 = Color(0x0DBB3D3D)
 
-val BusbyGreenLight = Color(0xFF7FF8BF)
+
+val BusbyGreenLight = Color(0xFFF9F9F9)
 val BusbyGreen30Light = Color(0x7F7FF8BF)
 val BusbyGreen10Light = Color(0x337FF8BF)
 val BusbyGreen5Light = Color(0x197FF8BF)
-val BusbyBlackLight = Color(0xFF808080)
-val BusbyGrayLight = Color(0xFFC4C4C4)
+val BusbyBlackLight = Color(0xFFffffff)
+val BusbyGrayLight = Color(0xFF999999)
 val BusbyGray40Light = Color(0x99C4C4C4)
 val BusbyDarkGrayLight = Color(0xFF545454)
-val BusbyWhiteLight = Color(0xFFFFFFFF)
+val BusbyWhiteLight = Color(0xFF000000)
 val BusbyDarkRedLight = Color(0xFFE68A8A)
 val BusbyDarkRed5Light = Color(0x26E68A8A)
