--- designsystem/src/main/java/me/androidbox/designsystem/ui/theme/Icons.kt
@@ -1,100 +1,0 @@
-package me.androidbox.designsystem.ui.theme
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.vectorResource
-import me.androidbox.designsystem.R
-
-val AnalyticIcon: ImageVector
-    @Composable
-    get() {
-        return ImageVector.vectorResource(id = R.drawable.analytics)
-    }
-
-val ArrowLeftIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.arrow_left)
-
-val ArrowRightIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.arrow_right)
-
-val CalendarIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.calendar)
-
-val CheckIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.check)
-
-val CrossIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.cross)
-
-val EmailIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.email)
-
-val ExclamationMarkIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.danger)
-
-val EyeClosedIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.eye_closed)
-
-val EyeOpenedIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.eye_opened)
-
-val FinishIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.finish)
-
-val KeyboardArrowDownIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.keyboard_arrow_down)
-
-val KeyboardArrowUpIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.keyboard_arrow_up)
-
-val LocationIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.location)
-
-val LockIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.lock)
-
-val LogoIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.logo)
-
-val LogoutIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.logout)
-
-val PauseIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.pause)
-
-val PersonIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.person)
-
-val RunOutlinedIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.run_outlined)
-
-val RunIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.run)
-
-val StartIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.start)
-
-val StopIcon: ImageVector
-    @Composable
-    get() = ImageVector.vectorResource(id = R.drawable.stop)
