+++ designsystem/src/main/java/me/androidbox/designsystem/ui/theme/Theme.kt
@@ -1,14 +1,10 @@
 package me.androidbox.designsystem.ui.theme
 
-import android.app.Activity
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.SideEffect
-import androidx.compose.ui.platform.LocalView
-import androidx.core.view.WindowCompat
-
 
 val DarkColorScheme = darkColorScheme(
     primary = BusbyGreen,
@@ -17,21 +13,21 @@ val DarkColorScheme = darkColorScheme(
     secondary = BusbyWhite,
     tertiary = BusbyWhite,
     primaryContainer = BusbyGreen30,
-    onPrimary = BusbyBlack,
+    onPrimary = BusbyBlackLight,
     onBackground = BusbyWhite,
     onSurface = BusbyWhite,
     onSurfaceVariant = BusbyGray,
     error = BusbyDarkRed
 )
 
-val LightColorScheme = darkColorScheme(
+val LightColorScheme = lightColorScheme(
     primary = BusbyGreenLight,
     background = BusbyBlackLight,
     surface = BusbyDarkGrayLight,
     secondary = BusbyWhiteLight,
     tertiary = BusbyWhiteLight,
     primaryContainer = BusbyGreen30Light,
-    onPrimary = BusbyBlackLight,
+    onPrimary = BusbyBlack,
     onBackground = BusbyWhiteLight,
     onSurface = BusbyWhiteLight,
     onSurfaceVariant = BusbyGrayLight,
@@ -39,32 +35,14 @@ val LightColorScheme = darkColorScheme(
 )
 
 @Composable
-fun LoginAppTheme(
+fun BusbyCoinsTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
-    // Dynamic color is available on Android 12+
-    dynamicColor: Boolean = true,
     content: @Composable () -> Unit
 ) {
-    /* val colorScheme = when {
-         dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
-             val context = LocalContext.current
-             if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
-         }
-
-         darkTheme -> DarkColorScheme
-         else -> LightColorScheme
-     }*/
-
-    val colorScheme = if(darkTheme) DarkColorScheme else LightColorScheme
-
-    val view = LocalView.current
-    if (!view.isInEditMode) {
-        SideEffect {
-            val window = (view.context as Activity).window
-            WindowCompat.getInsetsController(window, view).isAppearanceLightStatusBars = false
-        }
+    val colorScheme = when {
+        darkTheme -> DarkColorScheme
+        else -> LightColorScheme
     }
-
 
     MaterialTheme(
         colorScheme = colorScheme,
