+++ designsystem/src/main/java/me/androidbox/designsystem/ui/theme/Type.kt
@@ -2,66 +2,33 @@ package me.androidbox.designsystem.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
-import me.androidbox.designsystem.R
 
-val Poppins = FontFamily(
-    Font(
-        resId = R.font.poppins_light,
-        weight = FontWeight.Light
-    ),
-    Font(
-        resId = R.font.poppins_regular,
-        weight = FontWeight.Normal
-    ),
-    Font(
-        resId = R.font.poppins_medium,
-        weight = FontWeight.Medium
-    ),
-    Font(
-        resId = R.font.poppins_bold,
-        weight = FontWeight.Bold
-    ),
-    Font(
-        resId = R.font.poppins_semibold,
-        weight = FontWeight.SemiBold
-    )
-)
-
+// Set of Material typography styles to start with
 val Typography = Typography(
-    bodySmall = TextStyle(
-        fontFamily = Poppins,
-        fontWeight = FontWeight.Normal,
-        fontSize = 12.sp,
-        lineHeight = 20.sp,
-        color = BusbyGray
-    ),
-    bodyMedium = TextStyle(
-        fontFamily = Poppins,
-        fontWeight = FontWeight.Medium,
-        fontSize = 14.sp,
-        lineHeight = 22.sp,
-    ),
     bodyLarge = TextStyle(
-        fontFamily = Poppins,
+        fontFamily = FontFamily.Default,
         fontWeight = FontWeight.Normal,
         fontSize = 16.sp,
         lineHeight = 24.sp,
-        letterSpacing = 0.4.sp
-    ),
-    labelLarge = TextStyle(
-        fontFamily = Poppins,
+        letterSpacing = 0.5.sp
+    )
+    /* Other default text styles to override
+    titleLarge = TextStyle(
+        fontFamily = FontFamily.Default,
         fontWeight = FontWeight.Normal,
-        fontSize = 14.sp,
-        lineHeight = 24.sp,
+        fontSize = 22.sp,
+        lineHeight = 28.sp,
+        letterSpacing = 0.sp
     ),
-    headlineMedium = TextStyle(
-        fontFamily = Poppins,
-        fontWeight = FontWeight.SemiBold,
-        fontSize = 24.sp,
-        color = BusbyWhite
-    ),
+    labelSmall = TextStyle(
+        fontFamily = FontFamily.Default,
+        fontWeight = FontWeight.Medium,
+        fontSize = 11.sp,
+        lineHeight = 16.sp,
+        letterSpacing = 0.5.sp
+    )
+    */
 )
