+++ presentation/src/main/java/me/androidbox/presentation/coin_list/components/CoinDetailVerticalCard.kt
@@ -32,7 +32,7 @@ import io.kamel.image.KamelImage
 import io.kamel.image.asyncPainterResource
 import me.androidbox.presentation.R
 import me.androidbox.presentation.coin_list.CoinListState
-import me.androidbox.presentation.ui.theme.BusbyCoinsTheme
+import me.androidbox.designsystem.ui.theme.BusbyCoinsTheme
 
 @Composable
 fun CoinDetailVerticalCard(
